Cleaning up fixtures to be more resilient [PYTHON-3821]
